+++ src/components/spartan/SButtonGroup/SButtonGroupItem.vue
@@ -29,7 +29,7 @@ const iconClass = computed(() => [
         :class="
             twMerge([
                 SCN.focusRingPrimary,
-                'group group group relative inline-flex items-center justify-center gap-2 px-3 py-2 text-sm font-medium text-gray-900 outline outline-1 -outline-offset-1 outline-gray-300 focus:z-20 focus:outline-primary-300 active:bg-primary-50 active:text-primary-600',
+                'group relative inline-flex items-center justify-center gap-2 px-3 py-2 text-sm font-medium text-gray-900 outline outline-1 -outline-offset-1 outline-gray-300 focus:z-20 focus:outline-primary-300 active:bg-primary-50 active:text-primary-600',
                 active ? 'z-10 bg-primary-50 text-primary-600 outline-primary-300' : 'bg-white hover:bg-gray-50',
                 endIcon && 'flex-row-reverse',
                 disabled && 'pointer-events-none opacity-50',