+++ src/components/spartan/SInputDate/SInputDate.stories.ts
@@ -1,5 +1,5 @@
 import SInputDate from './SInputDate.vue';
-import { buildSourceBinding, createDefault, createHistory } from '@/helpers';
+import { buildSourceBinding, createDefault, createHistory, createVariation } from '@/helpers';
 import { ref } from 'vue';
 
 export default {
@@ -61,7 +61,7 @@ export const Default = createDefault({
         const value = ref(null);
         return { value };
     },
-    template: `<SInputDate v-model="value" v-bind="args" />`,
+    template: `<SInputDate v-model="value" v-bind="args" model-type="dd-MM-yyyy" /><pre>{{value}}</pre>`,
     transform: (args) => `<SInputDate ${sourceBinding(args)} />`,
     args: {
         hideInputIcon: false,
@@ -69,3 +69,15 @@ export const Default = createDefault({
         placeholder: 'Select a date',
     },
 });
+
+export const Base = createVariation({
+    components: { SInputDate },
+    setup: () => {
+        // const value = ref(null);
+        const value = ref('29-01-2000 14:30');
+        return { value };
+    },
+    containerClass: 'h-[500px] w-[300px]',
+    template: `<SInputDate v-model="value" v-bind="args" model-type="dd-MM-yyyy HH:mm" />
+<pre>{{value}}</pre>`,
+})