+++ src/components/spartan/SInputDateBlock/SInputDateBlock.vue
@@ -2,9 +2,10 @@
 import { SInputDate, type TInputDateProps, type TInputDateEmits } from '../SInputDate';
 import { BlockWrapper, type TBlockWrapperProps } from '@internal';
 import { extractWrapperProps } from '@/helpers';
+import type { VueDatePickerProps } from '@vuepic/vue-datepicker';
 
 defineEmits<TInputDateEmits>();
-const props = defineProps<Partial<TBlockWrapperProps> & TInputDateProps>();
+const props = defineProps<Partial<TBlockWrapperProps> & TInputDateProps & VueDatePickerProps>();
 const [blockWrapperProps, inputDateProps] = extractWrapperProps<TInputDateProps>(props);
 </script>